requirements.txt
streamlit
ultralytics
pillow
numpy
pandas
opencv-python-headless
imageio-ffmpeg